orchestration/requirements.txt
# Dagster Core
dagster==1.9.3
dagster-webserver==1.9.3
dagster-postgres==0.25.3

# Lineage & Observability (latest version for protobuf compatibility)
openlineage-python==1.9.1

# Data Processing
pandas==2.2.0
pyarrow==14.0.1
polars==0.20.0
duckdb==0.10.0

# Data Quality
great-expectations==0.18.19

# Database Connectors
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# S3/MinIO
boto3==1.34.10
minio==7.2.3

# HTTP/API
requests==2.31.0
httpx==0.26.0

# Weather API (backup: open-meteo)
openmeteo-requests==1.2.0

# Streaming (Kafka + WebSocket)
kafka-python==2.0.2
websockets==12.0

# Lakehouse (Apache Iceberg)
pyiceberg[pyarrow,s3fs]==0.6.1
requests-cache==1.1.1
retry-requests==2.0.0

# DuckDB for SQL queries on Iceberg
duckdb==1.1.3
duckdb-engine==0.13.2

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
